requirements.txt
flask==3.0.3
python-dotenv==1.0.1
web3==7.3.0
